automation/WebAutomation/requirements.txt
selenium~=4.33.0
pytest~=8.4.0
allure-python-commons~=2.15.0